refactor: getReportDetail 메서드를 SuspectController에서 ReportController로 이동
